test: fix Node-host Windows fixtures
Merge isolated spawn overrides case-insensitively so Node receives the fake-command PATH instead of an inherited Path entry. Reuse the existing helper and let the Windows timeout fixture retry cleanup while its killed command wrapper releases generated files.

Amp-Thread-ID: https://ampcode.com/threads/T-01a06c45-3cbe-743a-8d4e-ade43f745f88
