+++ tests/integrations/hook-helpers.ts
@@ -14,6 +14,7 @@ import { runHermesAgentHook } from '@/integrations/hermes-agent/hook';
 import type { InstallTargetChoice } from '@/integrations/install/choices';
 import type { InstallAction } from '@/integrations/install/targets';
 import { runKimiCodeHook } from '@/integrations/kimi-code/hook';
+import { createSpawnEnv } from '../helpers';
 
 /**
  * Shared test helpers for CLI hook integration tests.
@@ -360,24 +361,6 @@ export async function runHook(
   cwd = TEST_HOOK_CWD,
 ): Promise<HookResult> {
   return runCli(['hook', flag], input, env, cwd);
-}
-
-export function createSpawnEnv(overrides: Record<string, string>) {
-  const overriddenNames = new Set(
-    Object.keys(overrides).map((name) =>
-      process.platform === 'win32' ? name.toLowerCase() : name,
-    ),
-  );
-  return {
-    ...Object.fromEntries(
-      Object.entries(process.env).filter(
-        (entry): entry is [string, string] =>
-          entry[1] !== undefined &&
-          !overriddenNames.has(process.platform === 'win32' ? entry[0].toLowerCase() : entry[0]),
-      ),
-    ),
-    ...overrides,
-  };
 }
 
 export async function runCli(
